+++ src/Bpp/Qt/Tree/TreeCanvasControlers.cpp
@@ -201,7 +201,8 @@ void TreeCanvasControlers::treeDrawingUnitChanged()
   if (!treeCanvas_) return;
   if (! blockSignal_)
   {
-    treeCanvas_->setDrawingSize(widthCtrl_->value(), heightCtrl_->value());
+    treeCanvas_->setDrawingSize(static_cast<unsigned int>(widthCtrl_->value()),
+		                static_cast<unsigned int>(heightCtrl_->value()));
   }
 }
 
@@ -210,8 +211,8 @@ void TreeCanvasControlers::actualizeOptions()
   if (!treeCanvas_) return;
   AbstractDendrogramPlot* current = dynamic_cast<AbstractDendrogramPlot*>(treeCanvas_->getTreeDrawing());
   blockSignal_ = true; //Dirty trick but no choice!
-  widthCtrl_->setValue(treeCanvas_->drawingWidth());
-  heightCtrl_->setValue(treeCanvas_->drawingHeight());
+  widthCtrl_->setValue(static_cast<int>(treeCanvas_->drawingWidth()));
+  heightCtrl_->setValue(static_cast<int>(treeCanvas_->drawingHeight()));
   drawingCtrl_->setCurrentIndex(availableTreeDrawings_.indexOf(QString(current->getName().c_str())));
   if (current->getHorizontalOrientation() == AbstractDendrogramPlot::ORIENTATION_LEFT_TO_RIGHT)
     orientationLeftRight_->buttons()[0]->setChecked(true);
